requirements.txt

chainlit>=1.1.0
langchain>=0.2.10
langchain-core>=0.2.21
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
langchain-google-genai>=1.0.6
langchain-chroma>=0.1.1
chromadb>=0.5.4
pypdf>=4.3.1
PyMuPDF>=1.24.0   
Pillow>=10.2.0
python-dotenv>=1.0.1
